Ensure the dependencies are installed Rename main module to orchestrator
